Deployment/requirements.txt
tensorflow==2.16.1
opencv-python==4.8.1.78
numpy
streamlit
Pillow
